% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{calcul_distribution_semaine}
\alias{calcul_distribution_semaine}
\title{Calculer la distribution hebdomadaire des trajets}
\usage{
calcul_distribution_semaine(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant les donnees de trajets velo.}
}
\value{
Un data.frame contenant les jours de la semaine et le nombre de
trajets associe.
}
\description{
Cette fonction calcule la repartition du nombre total de trajets selon le
jour de la semaine.
}
\examples{
calcul_distribution_semaine(df_velo)
}
