% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{compter_nombre_boucle}
\alias{compter_nombre_boucle}
\title{Compter le nombre de boucles de comptage}
\usage{
compter_nombre_boucle(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant les donnees de trajets velo.}
}
\value{
Un entier correspondant au nombre de boucles distinctes.
}
\description{
Cette fonction calcule le nombre de boucles de comptage distinctes presentes
dans un jeu de donnees.
}
\examples{
compter_nombre_boucle(df_velo)
}
