% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtre_anomalie}
\alias{filtre_anomalie}
\title{Filtrer les anomalies dans les trajets velo}
\usage{
filtre_anomalie(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant les donnees de trajets velo.}
}
\value{
Un data.frame filtre.
}
\description{
Cette fonction filtre un jeu de donnees de trajets velo afin de conserver
uniquement les observations sans anomalie detectee, avec un total strictement
positif et inferieur a 10000.
}
\examples{
filtre_anomalie(df_velo)
}
