% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtrer_trajet}
\alias{filtrer_trajet}
\title{Filtrer un jeu de donnees par numero de boucle}
\usage{
filtrer_trajet(trajet, boucle)
}
\arguments{
\item{trajet}{Un data.frame contenant les donnees de trajets velo.}

\item{boucle}{Un vecteur de numeros de boucle a conserver.}
}
\value{
Un data.frame filtre sur les boucles selectionnees.
}
\description{
Cette fonction permet de selectionner uniquement certaines boucles de
comptage dans un jeu de donnees de trajets velo.
}
\examples{
filtrer_trajet(trajet = df_velo, boucle = c("880", "881"))
}
