% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{plot_distribution_semaine}
\alias{plot_distribution_semaine}
\title{Representer la distribution hebdomadaire des trajets}
\usage{
plot_distribution_semaine(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant les donnees de trajets velo.}
}
\value{
Un graphique \code{ggplot2}.
}
\description{
Cette fonction produit un graphique en barres de la distribution du nombre
total de trajets selon le jour de la semaine, apres filtrage des anomalies.
}
\examples{
plot_distribution_semaine(df_velo)
}
