% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{trouver_trajet_max}
\alias{trouver_trajet_max}
\title{Trouver le trajet avec le nombre maximal de passages}
\usage{
trouver_trajet_max(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant les donnees de trajets velo.}
}
\value{
Un data.frame contenant la boucle de comptage, le jour et le total
pour le ou les trajets maximaux.
}
\description{
Cette fonction renvoie la ou les observations pour lesquelles le nombre total
de trajets est maximal.
}
\examples{
trouver_trajet_max(df_velo)
}
