% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{calcul_distribution_semaine}
\alias{calcul_distribution_semaine}
\title{Calculer la distribution hebdomadaire des trajets}
\usage{
calcul_distribution_semaine(trajet, filtre = TRUE)
}
\arguments{
\item{trajet}{Un data.frame contenant les donnees de trajets velo.}

\item{filtre}{Un booléen indiquant si les anomalies doivent être filtrées
avant le calcul.}
}
\value{
Un data.frame contenant les jours de la semaine et le nombre de
trajets associé.
}
\description{
Cette fonction calcule la repartition du nombre total de trajets selon le
jour de la semaine.
}
\examples{
calcul_distribution_semaine(df_velo)
calcul_distribution_semaine(df_velo, filtre = FALSE)
}
